Refactor computePower function and Added test cases foe this class.

+++ src/com/epam/dsalgo/ComputePower.java
@@ -17,7 +17,7 @@ public class ComputePower {
         System.out.println(cp.power(a, n));
     }
 
-    public long power(int a, int n) {
+    public double power(int a, int n) {
         if (n <= 0) return 1;
         return a * power(a, n - 1);
     }
